refactored and some updates
