+++ src/Components/HomeSections/Benefits/Benefits.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/colors.scss";
+@import "../../../styles/colors.scss";
 
 .headline {
   color: $black;
