+++ src/Components/HomeSections/HomeHeader/HomeHeader.module.scss
@@ -1,7 +1,7 @@
-@import "../../styles/colors.scss";
-@import "../../styles/typography.scss";
-@import "../../styles/homeImages.scss";
-@import "../../styles/global.scss";
+@import "../../../styles/colors.scss";
+@import "../../../styles/typography.scss";
+@import "../../../styles/homeImages.scss";
+@import "../../../styles/global.scss";
 
 .row {
   justify-content: space-between;
