+++ src/Components/HomeSections/Upcoming/Upcoming.module.scss
@@ -1,5 +1,5 @@
-@import "../../styles/colors.scss";
-@import "../../styles/homeImages.scss";
+@import "../../../styles/colors.scss";
+@import "../../../styles/homeImages.scss";
 
 .headline {
   color: $black;
