+++ src/Components/HomeSections/Welcome/Welcome.module.scss
@@ -1,5 +1,5 @@
-@import "../../styles/homeImages.scss";
-@import "../../styles/colors.scss";
+@import "../../../styles/homeImages.scss";
+@import "../../../styles/colors.scss";
 
 .headline {
   color: $black;
